requirements.txt
aioredis==1.2.0
amqp==2.4.1
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
Babel==2.6.0
bcrypt==3.1.6
billiard==3.5.0.5
celery==4.2.1
celery-once==2.0.0
certifi==2018.11.29
cffi==1.12.0
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.5
daphne==2.2.5
Django==2.1.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-filters==0.2.1
django-simple-history==2.7.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
fabric2==2.4.0
flower==0.9.2
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
invoke==1.2.0
itypes==1.1.0
Jinja2==2.10
kombu==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.0
msgpack==0.6.1
msgpack-python==0.5.6
mysqlclient==1.3.13
nose==1.3.7
paramiko==2.4.2
paramiko-expect==0.2.8
Pillow==5.4.1
pyasn1==0.4.5
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
python-nmap==0.6.1
pytz==2018.9
PyYAML==3.13
redis==2.10.6
requests==2.21.0
ruamel.yaml==0.15.80
six==1.12.0
tornado==5.1.1
Twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
zope.interface==4.6.0
